@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 
 dependencies:
@@ -28,29 +28,29 @@ dependencies:
      sdk: flutter
   intl:
   #intl_translation:
-  flutter_keychain: ^1.0.0
-  flutter_secure_storage: ^3.3.5
-  uuid: 2.2.2
-  device_info: ^1.0.0
-  http: ^0.12.2
-  flutter_screenutil: ^4.0.2+1
-  email_validator: ^1.0.6
-  json_annotation: ^3.0.0
-  grouped_list: ^3.7.1
-  settings_ui: ^0.5.0
-  cupertino_icons: ^0.1.3
-  flutter_cupertino_settings: ^0.3.1
+  flutter_keychain: ^2.0.1
+  flutter_secure_storage: ^4.2.0
+  uuid: ^3.0.4
+  device_info: ^2.0.2
+  http: ^0.13.3
+  flutter_screenutil: ^5.0.0+2
+  email_validator: ^2.0.1
+  json_annotation: ^4.0.1
+  grouped_list: ^4.0.0
+  settings_ui: ^1.0.0
+  cupertino_icons: ^1.0.3
+  flutter_cupertino_settings: ^0.5.0
   #card_settings: ^1.16.2
-  flutter_colorpicker: ^0.3.5
-  flutter_speed_dial: ^1.2.5
+  flutter_colorpicker: ^0.4.0
+  flutter_speed_dial: ^3.0.5
   flutter_tags: ^0.4.9+1
   cupertino_radio_choice: ^1.0.1+1
-  modal_bottom_sheet: ^1.0.0+1
-  flutter_slidable: ^0.5.7
-  implicitly_animated_reorderable_list: ^0.3.2
+  modal_bottom_sheet: ^2.0.0
+  flutter_slidable: ^0.6.0
+  implicitly_animated_reorderable_list: ^0.4.0
   #flutter_material_pickers: ^1.9.1
-  reorderables: ^0.3.2
-  event_bus: ^1.1.1
+  reorderables: ^0.4.1
+  event_bus: ^2.0.0
   status_change: ^2.0.0
   webview_flutter: ^2.0.8
   flutter_youtube_view: ^2.0.0
@@ -64,11 +64,11 @@ dynamic_theme: ^1.0.1
 google_fonts: ^1.1.0
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-analyzer: ^0.41.2
+analyzer: ^1.7.1
 
 dev_dependencies:
-  build_runner: ^1.11.0
-  json_serializable: ^3.5.1
+  build_runner: ^2.0.5
+  json_serializable: ^4.1.0
 
 
   flutter_test:
