@@ -6,9 +6,9 @@ part 'SignupMailCnfResult.g.dart';
 // root flutter packages pub run build_runner build watch
 @JsonSerializable()
 class SignupMailCnfResult {
-  String statu;
-  String uuid;
-  List<Errs> errs;
+  String? statu;
+  String? uuid;
+  List<Errs>? errs;
 
   SignupMailCnfResult(this.statu, this.errs);
 
@@ -20,13 +20,19 @@ class SignupMailCnfResult {
 
 @JsonSerializable()
 class Errs{
-  String msg;
-  String errField;
+  String? msg;
+  String? errField;
 
   Errs(this.msg, this.errField);
 
-  factory Errs.fromJson(Map<String, dynamic> json) =>
-      _$ErrsFromJson(json);
+  factory Errs.fromJson(Map<String, dynamic>? json) {
+    if (json != null) {
+      return _$ErrsFromJson(json);
+    }
+    else {
+      return Errs("","");
+    }
+  }
 
   Map<String, dynamic> toJson() => _$ErrsToJson(this);
 
