@@ -6,16 +6,17 @@ import 'package:mbook_flutter/src/comm/api/api.dart';
 import 'package:mbook_flutter/src/comm/appbar.dart';
 import 'package:mbook_flutter/src/comm/consts.dart';
 import 'package:mbook_flutter/src/comm/global.dart';
-import 'package:mbook_flutter/src/comm/model/AdditionInfoManaWidget.dart';
+import 'package:mbook_flutter/src/comm/model/AdditionalInfo.dart';
+import 'package:mbook_flutter/src/comm/model/AdditionalMana.dart';
 import 'package:mbook_flutter/src/comm/model/ShopInfo.dart';
 import 'package:mbook_flutter/src/comm/widgets/Image.dart';
-import 'package:mbook_flutter/src/comm/widgets/fb_additioninfo_mana.dart';
+import 'package:mbook_flutter/src/comm/widgets/fb_addition_info_mana.dart';
 import 'package:mbook_flutter/src/mystore/my_addition_page.dart';
 import 'package:webview_flutter/webview_flutter.dart';
 
 class MyStoreInfoPage extends StatefulWidget {
-  final ShopInfo _shopInfo;
-  AdditionalMana addtionInfoMana = new AdditionalMana();
+  ShopInfo? _shopInfo;
+  AdditionalMana? addtionInfoMana;
 
   MyStoreInfoPage(this._shopInfo);
 
@@ -43,7 +44,7 @@ class _MyStoreInfoPageState extends State<MyStoreInfoPage> {
 
   SwitchItem _nowExpanded = SwitchItem.none;
 
-  WebViewController _controller;
+  late WebViewController _controller;
 
   @override
   Widget build(BuildContext context) {
@@ -54,13 +55,13 @@ class _MyStoreInfoPageState extends State<MyStoreInfoPage> {
     return Scaffold(
       key: _baseInfoscaffoldKey,
       appBar: AppBarView.appbar("Store info", true, canSave: true, onSave: () {
-        GlobalFun.showSnackBar(context, _baseInfoscaffoldKey, "  Saving...");
-        Api.saveMyShopInfo(context, widget._shopInfo).whenComplete(() {
+        GlobalFun.showSnackBar(context, _baseInfoscaffoldKey, null, "  Saving...");
+        Api.saveMyShopInfo(context, widget._shopInfo!).whenComplete(() {
           GlobalFun.removeCurrentSnackBar(_baseInfoscaffoldKey);
         }).catchError((e) {
-          GlobalFun.showSnackBar(context, _baseInfoscaffoldKey, e.toString());
+          GlobalFun.showSnackBar(context, _baseInfoscaffoldKey, null, e.toString());
         });
-      }),
+      }, context: context),
       body:
           //Scrollbar(child:SingleChildScrollView(child:
           DefaultTabController(
@@ -110,41 +111,41 @@ class _MyStoreInfoPageState extends State<MyStoreInfoPage> {
           GlobalFun.fbInputBox(
             context,
             "Store name",
-            widget._shopInfo.shopName,
+            widget._shopInfo?.shopName,
             (value) {
               setState(() {
-                widget._shopInfo.shopName = value;
+                widget._shopInfo?.shopName = value;
               });
             },
           ),
           GlobalFun.fbInputBox(
             context,
             "Tel",
-            widget._shopInfo.shopTel,
+            widget._shopInfo?.shopTel,
             (value) {
               setState(() {
-                widget._shopInfo.shopTel = value;
+                widget._shopInfo?.shopTel = value;
               });
             },
           ),
           GlobalFun.fbInputBox(
             context,
             "Locate",
-            widget._shopInfo.shopAddr,
+            widget._shopInfo?.shopAddr,
             (value) {
               setState(() {
-                widget._shopInfo.shopAddr = value;
+                widget._shopInfo?.shopAddr = value;
               });
             },
           ),
-          GlobalFun.fbInputBox(context, "Image", widget._shopInfo.shopPicUrl,
+          GlobalFun.fbInputBox(context, "Image", widget._shopInfo?.shopPicUrl,
               (value) {
             setState(() {
-              widget._shopInfo.shopPicUrl = value;
+              widget._shopInfo?.shopPicUrl = value;
             });
           },
               valueWidget: Row(children: [
-                Flexible(child: new MBImage(url: widget._shopInfo.shopPicUrl))
+                Flexible(child: new MBImage(url: widget._shopInfo?.shopPicUrl))
               ])),
         ]));
   }
@@ -153,7 +154,7 @@ class _MyStoreInfoPageState extends State<MyStoreInfoPage> {
     return MyAdditionPage(addtionInfoMana: widget.addtionInfoMana);
   }
 
-  ScrollController scrollController;
+  late ScrollController scrollController;
   bool dialVisible = true;
 
   void setDialVisible(bool value) {
@@ -166,7 +167,7 @@ class _MyStoreInfoPageState extends State<MyStoreInfoPage> {
 class TabInfo {
   const TabInfo({this.title, this.icon, this.widget});
 
-  final String title;
-  final IconData icon;
-  final Widget widget;
+  final String? title;
+  final IconData? icon;
+  final Widget? widget;
 }
