@@ -7,7 +7,7 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,32 +20,46 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'zh_CN';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "common_all_title" : MessageLookupByLibrary.simpleMessage("Food BooK"),
-    "home_login_button_title" : MessageLookupByLibrary.simpleMessage("Join"),
-    "home_scan_button_title" : MessageLookupByLibrary.simpleMessage("Scan"),
-    "login_email_validator_empty_msg" : MessageLookupByLibrary.simpleMessage("Please enter Your email address."),
-    "login_email_validator_not_valid_msg" : MessageLookupByLibrary.simpleMessage("Please enter a valid email address. "),
-    "login_login_title" : MessageLookupByLibrary.simpleMessage("Sign in"),
-    "login_mail_hintText" : MessageLookupByLibrary.simpleMessage("Mail"),
-    "login_password_hintText" : MessageLookupByLibrary.simpleMessage("Password"),
-    "login_password_validator_empty_msg" : MessageLookupByLibrary.simpleMessage("Please enter your password. "),
-    "login_title" : MessageLookupByLibrary.simpleMessage("Sign in"),
-    "menu_login_title" : MessageLookupByLibrary.simpleMessage("Sign in"),
-    "menu_logout_title" : MessageLookupByLibrary.simpleMessage("Sign out"),
-    "menu_mystore_title" : MessageLookupByLibrary.simpleMessage("My store"),
-    "menu_signup_title" : MessageLookupByLibrary.simpleMessage("Sign up"),
-    "mystore_title" : MessageLookupByLibrary.simpleMessage("My store"),
-    "signup_button_sending" : MessageLookupByLibrary.simpleMessage("Send"),
-    "signup_code_button_sending" : MessageLookupByLibrary.simpleMessage("Confirm"),
-    "signup_password_validator_cofict_msg" : MessageLookupByLibrary.simpleMessage("Inconsistent passwords"),
-    "signup_signup_title" : MessageLookupByLibrary.simpleMessage("Sign up"),
-    "signup_sigup_button" : MessageLookupByLibrary.simpleMessage("Sigup"),
-    "sigup_email_validator_empty_msg" : MessageLookupByLibrary.simpleMessage("Please enter Your email address."),
-    "sigup_email_validator_not_valid_msg" : MessageLookupByLibrary.simpleMessage("Please enter a valid email address. "),
-    "sigup_mail_cfn_hintText" : MessageLookupByLibrary.simpleMessage("Please enter the verification code"),
-    "sigup_mail_hintText" : MessageLookupByLibrary.simpleMessage("Mail"),
-    "sigup_password_hintText" : MessageLookupByLibrary.simpleMessage("Password"),
-    "sigup_password_validator_empty_msg" : MessageLookupByLibrary.simpleMessage("Please enter password. ")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "common_all_title": MessageLookupByLibrary.simpleMessage("Food BooK"),
+        "home_login_button_title": MessageLookupByLibrary.simpleMessage("Join"),
+        "home_scan_button_title": MessageLookupByLibrary.simpleMessage("Scan"),
+        "login_email_validator_empty_msg": MessageLookupByLibrary.simpleMessage(
+            "Please enter Your email address."),
+        "login_email_validator_not_valid_msg":
+            MessageLookupByLibrary.simpleMessage(
+                "Please enter a valid email address. "),
+        "login_login_title": MessageLookupByLibrary.simpleMessage("Sign in"),
+        "login_mail_hintText": MessageLookupByLibrary.simpleMessage("Mail"),
+        "login_password_hintText":
+            MessageLookupByLibrary.simpleMessage("Password"),
+        "login_password_validator_empty_msg":
+            MessageLookupByLibrary.simpleMessage(
+                "Please enter your password. "),
+        "login_title": MessageLookupByLibrary.simpleMessage("Sign in"),
+        "menu_login_title": MessageLookupByLibrary.simpleMessage("Sign in"),
+        "menu_logout_title": MessageLookupByLibrary.simpleMessage("Sign out"),
+        "menu_mystore_title": MessageLookupByLibrary.simpleMessage("My store"),
+        "menu_signup_title": MessageLookupByLibrary.simpleMessage("Sign up"),
+        "mystore_title": MessageLookupByLibrary.simpleMessage("My store"),
+        "signup_button_sending": MessageLookupByLibrary.simpleMessage("Send"),
+        "signup_code_button_sending":
+            MessageLookupByLibrary.simpleMessage("Confirm"),
+        "signup_password_validator_cofict_msg":
+            MessageLookupByLibrary.simpleMessage("Inconsistent passwords"),
+        "signup_signup_title": MessageLookupByLibrary.simpleMessage("Sign up"),
+        "signup_sigup_button": MessageLookupByLibrary.simpleMessage("Sigup"),
+        "sigup_email_validator_empty_msg": MessageLookupByLibrary.simpleMessage(
+            "Please enter Your email address."),
+        "sigup_email_validator_not_valid_msg":
+            MessageLookupByLibrary.simpleMessage(
+                "Please enter a valid email address. "),
+        "sigup_mail_cfn_hintText": MessageLookupByLibrary.simpleMessage(
+            "Please enter the verification code"),
+        "sigup_mail_hintText": MessageLookupByLibrary.simpleMessage("Mail"),
+        "sigup_password_hintText":
+            MessageLookupByLibrary.simpleMessage("Password"),
+        "sigup_password_validator_empty_msg":
+            MessageLookupByLibrary.simpleMessage("Please enter password. ")
+      };
 }
