@@ -8,11 +8,11 @@ part of 'TagInfo.dart';
 
 TagInfo _$TagInfoFromJson(Map<String, dynamic> json) {
   return TagInfo(
-    id: json['id'] as int,
-    data: json['data'] as String,
-    desc: json['desc'] as String,
-    propertyString: json['propertyString'] as String,
-  )..orders = json['orders'] as int;
+    id: json['id'] as int?,
+    data: json['data'] as String?,
+    desc: json['desc'] as String?,
+    propertyString: json['propertyString'] as String?,
+  )..orders = json['orders'] as int?;
 }
 
 Map<String, dynamic> _$TagInfoToJson(TagInfo instance) => <String, dynamic>{
