@@ -13,44 +13,44 @@ TextWidgetProperty _$TextWidgetPropertyFromJson(Map<String, dynamic> json) {
         const CustomColorConverter().fromJson(json['textColor'] as String),
     backColor:
         const CustomColorConverter().fromJson(json['backColor'] as String),
-    fontSize: (json['fontSize'] as num)?.toDouble(),
+    fontSize: (json['fontSize'] as num).toDouble(),
     fontWeight: json['fontWeight'] as int,
     italic: json['italic'] as bool,
-    letterSpacing: (json['letterSpacing'] as num)?.toDouble(),
-    paddingLeft: (json['paddingLeft'] as num)?.toDouble(),
-    paddingTop: (json['paddingTop'] as num)?.toDouble(),
-    paddingRight: (json['paddingRight'] as num)?.toDouble(),
-    paddingBottom: (json['paddingBottom'] as num)?.toDouble(),
+    letterSpacing: (json['letterSpacing'] as num).toDouble(),
+    paddingLeft: (json['paddingLeft'] as num).toDouble(),
+    paddingTop: (json['paddingTop'] as num).toDouble(),
+    paddingRight: (json['paddingRight'] as num).toDouble(),
+    paddingBottom: (json['paddingBottom'] as num).toDouble(),
     textTextAlign: json['textTextAlign'] as String,
-    borderWidth: (json['borderWidth'] as num)?.toDouble(),
+    borderWidth: (json['borderWidth'] as num).toDouble(),
     borderColor:
         const CustomColorConverter().fromJson(json['borderColor'] as String),
-    borderRadiusTopLeft: (json['borderRadiusTopLeft'] as num)?.toDouble(),
-    borderRadiusTopRight: (json['borderRadiusTopRight'] as num)?.toDouble(),
-    borderRadiusBottomLeft: (json['borderRadiusBottomLeft'] as num)?.toDouble(),
+    borderRadiusTopLeft: (json['borderRadiusTopLeft'] as num).toDouble(),
+    borderRadiusTopRight: (json['borderRadiusTopRight'] as num).toDouble(),
+    borderRadiusBottomLeft: (json['borderRadiusBottomLeft'] as num).toDouble(),
     borderRadiusBottomRight:
-        (json['borderRadiusBottomRight'] as num)?.toDouble(),
+        (json['borderRadiusBottomRight'] as num).toDouble(),
     shadowColor:
         const CustomColorConverter().fromJson(json['shadowColor'] as String),
-    shadowOffsetX: (json['shadowOffsetX'] as num)?.toDouble(),
-    shadowOffsetY: (json['shadowOffsetY'] as num)?.toDouble(),
-    shadowBlurRadius: (json['shadowBlurRadius'] as num)?.toDouble(),
-    shadowSpreadRadius: (json['shadowSpreadRadius'] as num)?.toDouble(),
-    minWidth: (json['minWidth'] as num)?.toDouble(),
+    shadowOffsetX: (json['shadowOffsetX'] as num).toDouble(),
+    shadowOffsetY: (json['shadowOffsetY'] as num).toDouble(),
+    shadowBlurRadius: (json['shadowBlurRadius'] as num).toDouble(),
+    shadowSpreadRadius: (json['shadowSpreadRadius'] as num).toDouble(),
+    minWidth: (json['minWidth'] as num).toDouble(),
     minWidthUnit: json['minWidthUnit'] as String,
-    minHeight: (json['minHeight'] as num)?.toDouble(),
+    minHeight: (json['minHeight'] as num).toDouble(),
     minHeightUnit: json['minHeightUnit'] as String,
-    maxWidth: (json['maxWidth'] as num)?.toDouble(),
-    maxHeight: (json['maxHeight'] as num)?.toDouble(),
+    maxWidth: (json['maxWidth'] as num).toDouble(),
+    maxHeight: (json['maxHeight'] as num).toDouble(),
     alignment: json['alignment'] as String,
     backalignment: json['backalignment'] as String,
-    marginLeft: (json['marginLeft'] as num)?.toDouble(),
-    marginTop: (json['marginTop'] as num)?.toDouble(),
-    marginRight: (json['marginRight'] as num)?.toDouble(),
-    marginBottom: (json['marginBottom'] as num)?.toDouble(),
+    marginLeft: (json['marginLeft'] as num).toDouble(),
+    marginTop: (json['marginTop'] as num).toDouble(),
+    marginRight: (json['marginRight'] as num).toDouble(),
+    marginBottom: (json['marginBottom'] as num).toDouble(),
     backImg: json['backImg'] as String,
-    spacingHWidth: (json['spacingHWidth'] as num)?.toDouble(),
-    spacingVWidth: (json['spacingVWidth'] as num)?.toDouble(),
+    spacingHWidth: (json['spacingHWidth'] as num).toDouble(),
+    spacingVWidth: (json['spacingVWidth'] as num).toDouble(),
     spacingColor:
         const CustomColorConverter().fromJson(json['spacingColor'] as String),
   );
